Move layout types to layout.go
